refactor(PT-51): reorganize components and remove unused/empty ones
