--- src/components/Profile/AddPetForm.tsx
@@ -1,5 +1,0 @@
-const AddPetForm = () => {
-  return <div></div>;
-};
-
-export default AddPetForm;
