--- src/components/Profile/EditProfileForm.tsx
@@ -1,5 +1,0 @@
-const EditProfileForm = () => {
-  return <div></div>;
-};
-
-export default EditProfileForm;
