--- src/components/Profile/MyBookings.tsx
@@ -1,5 +1,0 @@
-const MyBookings = () => {
-  return <div></div>;
-};
-
-export default MyBookings;
