--- src/components/Profile/PetList.tsx
@@ -1,5 +1,0 @@
-const PetList = () => {
-  return <div></div>;
-};
-
-export default PetList;
