--- src/components/Profile/ProfileInfo.tsx
@@ -1,5 +1,0 @@
-const ProfileInfo = () => {
-  return <div></div>;
-};
-
-export default ProfileInfo;
