--- src/components/Shared/StarsRating.tsx
@@ -1,5 +1,0 @@
-const StarsRating = () => {
-  return <div></div>;
-};
-
-export default StarsRating;
